WIP: mod: resource/csdk/connectivity/src/cainterfacecontroller.c (master)

+++ resource/csdk/connectivity/src/cainterfacecontroller.c
@@ -53,7 +53,7 @@
 
 static CAConnectivityHandler_t *g_adapterHandler = NULL;
 
-static uint32_t g_numberOfAdapters = 0;
+static size_t g_numberOfAdapters = 0;
 
 static CANetworkPacketReceivedCallback g_networkPacketReceivedCallback = NULL;
